Extract track in music listings to its own component, success check mark
