+++ src/main/js/src/components/Queue.js
@@ -6,7 +6,7 @@ function Queue({music}) {
             <h3 className="column-header">Queue</h3>
             <ul className="queue">
             {music && music.length > 0 ? music.map((track, index) => (
-                <li key={index * -1} className="list-group-item">
+                <li key={index * -1} className="list-group-item track">
                     <div>
                         {track.title} <br/>
                         <small className="text-muted">{track.artist} &middot; {track.album}</small><br/>
