+++ src/main/js/src/scss/components/_track.scss
@@ -1,0 +1,13 @@
+li.track {
+
+  img { // add to queue button
+    padding: 0 10px 0 10px;
+    width: 2.5rem;
+    border-radius: 10px;
+
+    &.button:hover {
+      cursor: pointer;
+      fill: red;
+    }
+  }
+}
