+++ src/main/js/src/scss/components/all.scss
@@ -1,6 +1,7 @@
 @import './header';
 @import './button';
-@import './available-music';
+@import './track-list';
+@import './track';
 @import './queue';
 @import './now-playing';
 @import './filter';
